astype

+++ src/luigi/Version2/featureEngineering.py
@@ -88,7 +88,7 @@ class featureEngineering(PostgresQuery):
         for row in df2:
             tupla = [tuple(x) for x in df2.values]
             print(tupla)
-            query = """INSERT INTO {} VALUES {};""".format(self.table,tupla[row])
+            query = """INSERT INTO {} VALUES {};""".format(self.table,tupla[row].astype(int))
             sql = query
             cursor.execute(sql)
